search/requirements.txt
PyWavelets==1.8.0
imagehash==4.3.2
numpy==2.2.6
pillow==11.2.1
scipy==1.15.3
opencv-python==4.11.0.86
pymongo==4.13.0
pyyaml==6.0.2
decorator==4.4.2
moviepy==1.0.3
lazy-loader==0.4 
networkx==3.5 
packaging==25.0 
scikit-image==0.25.2 
tifffile==2025.6.1
pynvml
pymongo
tqdm